Bug 732114 - part 1: inherit from AndroidSyncTestCase.
