+++ test/src/org/mozilla/android/sync/test/AndroidBrowserRepositoryTest.java
@@ -25,7 +25,6 @@ import org.mozilla.android.sync.test.helpers.ExpectStoreCompletedDelegate;
 import org.mozilla.android.sync.test.helpers.ExpectStoredDelegate;
 import org.mozilla.android.sync.test.helpers.WaitHelper;
 import org.mozilla.gecko.db.BrowserContract;
-import org.mozilla.gecko.sync.StubActivity;
 import org.mozilla.gecko.sync.Utils;
 import org.mozilla.gecko.sync.repositories.InactiveSessionException;
 import org.mozilla.gecko.sync.repositories.InvalidSessionTransitionException;
@@ -40,45 +39,12 @@ import org.mozilla.gecko.sync.repositories.domain.Record;
 
 import android.content.ContentValues;
 import android.content.Context;
-import android.test.ActivityInstrumentationTestCase2;
 import android.util.Log;
 
-public abstract class AndroidBrowserRepositoryTest extends ActivityInstrumentationTestCase2<StubActivity> {
-  
+public abstract class AndroidBrowserRepositoryTest extends AndroidSyncTestCase {
+
   protected AndroidBrowserRepositoryDataAccessor helper;
   protected static final String tag = "AndroidBrowserRepositoryTest";
-  
-  public AndroidBrowserRepositoryTest() {
-    super(StubActivity.class);
-  }
-
-  public Context getApplicationContext() {
-    return this.getInstrumentation().getTargetContext().getApplicationContext();
-  }
-
-  protected static void performWait(Runnable runnable) throws AssertionFailedError {
-    WaitHelper.getTestWaiter().performWait(runnable);
-  }
-  
-  protected static void performNotify() {
-    WaitHelper.getTestWaiter().performNotify();
-  }
-
-  protected static void performNotify(AssertionFailedError e) {
-    WaitHelper.getTestWaiter().performNotify(e);
-  }
-
-  protected static void performNotify(InactiveSessionException e) {
-    AssertionFailedError er = new AssertionFailedError("Inactive session.");
-    er.initCause(e);
-    WaitHelper.getTestWaiter().performNotify(er);
-  }
-  
-  protected static void performNotify(InvalidSessionTransitionException e) {
-    AssertionFailedError er = new AssertionFailedError("Invalid session transition.");
-    er.initCause(e);
-    WaitHelper.getTestWaiter().performNotify(er);
-  }
 
   protected void wipe() {
     if (helper == null) {
